Split link_ids into relocs & gctags
The previous implementation crashed once you loaded a single package
and then tried to link against two blobs. The issue turned out to be
that the `link_ids` are written in the order in which they are
encountered, but when saved to the sysimg they are split into separate
`gctags` and `relocs` groups. This effectively shuffles the `link_ids`
list and causes the wrong blob to be used for relocation.

This fixes the problem, but the harder work of diagnosing the problem
was done by vchuravy.

Co-authored-by:  Valentin Churavy <vchuravy@csail.mit.edu>